rewrite chapter tab
The chapter tab is an instance of the Tab-module as well. It is added
to the tab-registry.
Each chapter tab gets a reference to the update function to enable
highlighting of the currently played chapter.

TODO:  There is a need for a global update hook that can be listened to
by each tab.