juggle methods and properties between player and app
Add property player.defaults with the webPlayer defaults.
Deeplink parsing is part of the app callback (and still WIP).
The error handler that tries to remedy the problem by removing media
files that might be unrecognised is still part of the player module as
the list of all registered players.