enable timeline to end at a specific time
Timeline instances have a new property endTime - a timestamp.
Whenever the currentTime reaches or passes the endTime on update the player will
be stopped.
The endTime can be set with Timeline.stopAt(time).
The convenience method Timeline.playRange([start,end]) can be used to
share arbitrary parts of an episode.

Add tests for timeline.playRange method.