refactor progressbar module
Add ProgressBar#setProgress that takes a time as an argument and
renders the progress bar accordingly
It  is called from _update and the when the user clicks on the progress
bar or drags the slider.
No threshold on drag (was 0.02). It leads to flickering on drag while
playing
and prevents dragging the slider to the very start.