Changes for Handle in Progressbar
The handle has got hover effects now: When hovering the progressbar
element, the handle will transform from a thin line up to a square.
When hovering the handle itself, it will scale up a bit.

The handle will be displayed with a correct offset value - leaving half
of the handle visible when starting or reaching the end in the progress
bar.