feat (localization): set locale by configuration parameter
If “languageCode” is set to the name of an existing key in the locales
map the app will use this to translate all labels.
Currently, “de” and “en” are present.
“en” is used as a fallback.
All language files are required on module startup to get rid of extra
requests and to speed up initial loading.
A warning is written to the console when no translation is found for
the given languageCode.
On player initialisation the locale is set.