feat (moderator): no jQuery triggers embed mode
The moderator assumes that it runs in embed mode when jQuery is not
present.
It will then not replace the elements it found with the query but place
the corresponding iframe right next to it.

The old `podlovewebplayer()` API is there, if jQuery is present.

All functions were stripped from there dependendency to jQuery, which
was not that much luckily.

In `/examples/embed/friendly.html` you can see an example.

NOTE: Now lots of testing regarding cross-browser compatibility and
restrictions to cross-origin requests have to be performed.