fix(layout): One column in info tab on small screens
The column width for links and headlines in the show info was too small
for a decent display and readability.
For small screens up to a max width of 768px, only one column will be
displayed - the show info will be displayed below the episode info.
On larger screens, the 2 column display will be kept as before.